Beginning of migration: automatic tests pass, coverage is lousy, haven't checked the example site yet

+++ clean_build.sh
@@ -1,6 +1,12 @@
 #!/bin/bash
 
-git tag `grep "version" setup.py | cut -d "'" -f 2`
+version=`grep "__version__ = " yacon/__init__.py | cut -d "'" -f 2`
+
+git tag "$version"
+
+if [ "$?" != "0" ] ; then
+    exit $?
+fi
 
 rm -rf build
 rm -rf dist
@@ -8,6 +14,8 @@ python setup.py sdist
 python setup.py bdist_wheel
 
 echo "------------------------"
+echo 
+echo "Built version: $version"
 echo
 echo "now do:"
 echo "   twine upload dist/*"
